+++ src/kaleidoscope/Key/Mouse.h
@@ -8,8 +8,10 @@
 namespace kaleidoscope {
 
 class Key::Mouse {
+
  private:
-  uint16_t keycode_ : 8, type_id_ : 8;
+  byte keycode_ : 8, type_id_ : 8;
+
  public:
   byte keycode() const { return keycode_; }
